requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-shell
    #   rossum
    #   typer
    #   uvicorn
click-shell==2.1
    # via rossum
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
et-xmlfile==1.1.0
    # via openpyxl
fastapi==0.111.1
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
gunicorn==22.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   -r requirements.in
    #   fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via fastapi
jmespath==1.0.1
    # via rossum
jsondiff==2.2.0
    # via rossum
lxml==5.2.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via rossum
numpy==2.0.1
    # via pandas
openpyxl==3.0.10
    # via rossum
packaging==24.1
    # via
    #   gunicorn
    #   pytest
pandas==2.2.2
    # via rossum
pluggy==1.5.0
    # via pytest
polling2==0.5.0
    # via rossum
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
pytest==8.3.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.8
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   jsondiff
    #   uvicorn
redis==5.0.7
    # via -r requirements.in
requests==2.32.3
    # via rossum
rich==13.7.1
    # via typer
rossum==3.18.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tabulate==0.9.0
    # via rossum
tenacity==8.5.0
    # via rossum
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
uvicorn[standard]==0.30.3
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
xlrd==2.0.1
    # via rossum
